simplified serial.zig
Renamed a bunch of things that contained "Serial" and "Port" as it's redundant.

Renamed lists/list info to iterator/listIterator for clarity.

Added init func to create file and configure all in one go.

Removed baud maps in favor of std.c.speed_t, so most os should be covered.